+++ src/formatter/concerns/mod.rs
@@ -1,20 +1,3 @@
-/// Escape whatever's being sent
-/// in here to an ansi code
-pub struct Ansi {}
+mod ansi;
 
-impl Ansi {
-    /// Add the required escape and terminator characters to
-    /// an ansi code.
-    pub fn escape(code: u8) -> String {
-        format!("\x1B[{}m", code)
-    }
-}
-
-pub trait FromKey {
-    /// Define your own implementation of how
-    /// a given key will return the type that
-    /// implements it. A wrapper around FromStr
-    /// in case the key returns different things based
-    /// on some internal value.
-    fn from_key(key: &str) -> Option<String>;
-}
+pub use ansi::Ansi;
