+++ src/formatter/keys/mod.rs
@@ -1,5 +1,7 @@
-pub mod key;
-pub mod key_list;
+mod key;
+mod key_list;
+mod from_key;
 
 pub use key::Key;
 pub use key_list::KeyList;
+pub use from_key::FromKey;
